RT-1.5: Updating neighbor configuration for missing routing-policy and specific afi-safi config
This PR introduces 2 changes:

1. route policy is always added and configured on a per-neighbor basis if RoutePolicyUnderNeighborAfiSafi deviation is not enabled,

2. explicitly enables/disables both address families for each BGP neighbor (ipv4 en. / ipv6 dis. for v4 peers, v4 dis., v6 en. for v6 peers). This change is needed to align with the inheritance model discussed in Remove the default of "enabled" of BGP AFI-SAFI to support BGP inheritance model public#774 -> in the current code both IPv4 and IPv6 are enabled on a protocol level, so in a per-neighbor config we want to disable inheritance of the other protocol family.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."