OC-1.2: Update to support ExplicitPortSpeed & ExplicitInterfaceInDefaultVRF deviations
In this pr we introduce two changes:

1. when ExplicitInterfaceInDefaultVRF is used, we need to unassign interfaces from the default vrf before they can be assigned to a custom vrf.
unassignPort() function is added for this purpose, which is called only when ExplicitInterfaceInDefaultVRF is true
2. added fptest.SetPortSpeed() function calls when ExplicitPortSpeed is true

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."