RT-3.2: add missing instances of InterfaceEnabled deviation
This PR adds missing instances of InterfaceEnabled deviation:

1. ipv6 subif config in ATE version of the test
2. interface, ipv4 subif, ipv6 subif in the OTG version of the test

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."