gNOI-3.2: add check on linecard emptiness
Current version of the test selects the last removable linecard to reboot, however it ignores the empty flag (so it tries to reboot an unpopulated card in some cases). This PR introduces an additional check to skip cards that are not populated

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."