RT-2.1 and RT-2.2, Enabling "MissingValueForDefaults" deviation for default-value checks after adjacency
All the default values are checked via MissingValueForDefaults deviation before the adjacency,
Same is required for expected unchanged counters after adjacency checks.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."