RT-5.1: Enabling deviation ExplicitInterfaceInDefaultVRF and ExplicitPortSpeed
There are 2 changes -
 1. Adding deviations ExplicitInterfaceInDefaultVRF and  ExplicitPortSpeed
 2. Fixing mac-address to upper-case

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."