TE-11.3: Enabling deviation, updating NH index to integer and adding network-instance for DECAP lookup (#1246)
There are 2 changes made here -
   1. Enabling following deviations -
         - ExplicitInterfaceInDefaultVRF
         - ExplicitPortSpeed
         - ExplicitGRIBIUnderNetworkInstance
   2. Updating static-route NextHop index to an integer
   
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."