Revert "gnmi-1.10: check sw versions on supervisors instead of CPU"
This reverts commit b494d327f32ce7ec8d16887ac9044ecf3558e922.



This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.