GNMI-1.11 Updating the deviation config order
Moving ExplicitPortSpeed configuration after ExplicitInterfaceInDefaultVRF to avoid intermittent failure.
Sleep call under ExplicitPortSpeed gives enough time for subinterfaces to get updated in default-vrf,
Otherwise intermittent failure was observed.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."