RT-2.1: Adding deviation for multi-topology and restart-suppress
   1. Introducing devation ISISMultiTopologyUnsupported and ISISRestartSuppressUnsupported
   2. Skiping leaf check for multi-topology and restart-suppress when these deviations are marked true.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."