Pin openconfig/public repo when validating paths using a bash file.
This seems the cleanest way to do this. While having all the
dependencies in `go.mod` would be nice, it is not possible to have the
`require` in `go.mod` without making openconfig/public have a `go.mod`
file, and even if we could do that, `go mod vendor` will download
all dependencies for the module, and that doesn't seem right just to
maintain a view of a non-Go repo.