RT-2.1 and RT-2.2: OTG only changes based on PR-1229
These changes are specific to OTG to match exact changes done for ATE using PR-1229

Following changes are clubbed together as they are interdependent -

    1. As interface must be part of VRF, before it is attached to ISIS -
        - Moving deviations.ExplicitInterfaceInDefaultVRF config block from PushAndStart to PushDUT.
        - Passing testing.TB handle from PushAndStart to PushDUT
        - Updating PushDUT calls to pass testing.TB handle

    2. Updating interface-name usage in all the places based on deviations.ExplicitInterfaceInDefaultVRF
    3. Explicitly configuring Global.LevelCapability to Level_2, considering different vendor have different default.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."