te-3.1: added support for no default vrf ,MAC NH
1. added support for non-default VRF VRF-1 and moved the route to it
2. Added MAC next-hop coverage
3. Added MAC validation in MAC next-hop scenario

This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an as is basis without any warranties of any kind.