TE-11.3  implemented changes as per PR 1158
1. Added port 4 to topology and modified vrf names
2. Added port 1 to VRF-1
2. Added Decap and Encap support
3. Added backup support primary to decap+encap and decap+encap to decap
4. Added packet validation based on path
5. Added lookup vrf support in AddNH funtion and also optimized the code
This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.