GNMI-1.2: Adding deviation for authentication and updating ISIS configuration
Following changes are made -
1. Updating configuration to use existing deviation ISISprotocolEnabledNotRequired
2. Adding subinterface to default-NI using deviation ExplicitInterfaceInDefaultVRF
3. Introducing deviation for ISISGlobalAuthenticationNotRequired to avoid Global.Authentication-check for ISIS
4. Introducing deviation for ISISLevelAuthenticationNotRequired to avoid Level Authentication configuration for ISIS
5. Enabling ISIS ipv4-unicast at global config.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."