GNMI-1.3: Updating configuration for routing-policy under peer-group and interface to default-ni
1. Updating configuraion to add interface to Default NI using deviation ExplicitInterfaceInDefaultVRF
2. Adding ExplicitPortSpeed deviation config
3. Adding explicit configuration for routing-policy ALLOW under peer-group
4. Updating AS-Path sequence policy to ACCEPT_ROUTE explicitly while it updates AS-Seq

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."