Revert "te-14.1: use gnmi replace in createSubifDUT()"
This reverts commit 5e83ba81263c956af18d668de64cc8e0a775dc9f.



This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.