RT-5.4: Enabling deviation and updating config order
Following changes are made -
1. Enabling deviation ExplicitInterfaceInDefaultVRF and ExplicitPortSpeed
2. Moving ATE configuration before verifyDUT is called to make sure interfaces are up
3. Adding batch-delete under clearAggregate to delete dependent configs together

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."