RT-1.4 Adding Native SRL ACL compatibility (#1389)
* Added compatibility to configure ACLs via Native SRL model as well as a deviation to ensure no other vendor is impacted.

This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.