Change `make openconfig_public` to update in lock-step with new OC releases. (#1408)
I don't think it's necessary to require manual intervention when it comes to updating major OC releases. We can start to provide failure signals as soon as a new version of OC has been released. Affected PRs will be given signal to update themselves, and repo maintainers are given a signal to start fix existing tests.