RT-3.2: added check for ARP resolution from OTG (#1638)
* adding case insenitive mac address check

This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an as is basis without any warranties of any kind.

* Revert "adding case insenitive mac address check"

This reverts commit 8c13da0561b270ca02660108f0b2c4e43653e6e2.

* added check for ARP resolution from OTG

This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an as is basis without any warranties of any kind

* repalcing waitOTGARPENtry with existing fucntion WaitForARP