fixed error in PPS set and added enhacement
1. added PPS in flow creation to have accurate loss percentage
calcuation
2. added support to have multiple flows to make sure that traffic is
3. replaced waitToGARP code with util function

Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.