TE-3.1: Updating metadata.textproto with missing deviation (#1746)
* TE-3.1: Updating metadata.textproto with missing deviation

Adding following 2 missing deviations to metadata.textproto -
    explicit_gribi_under_network_instance: true
    explicit_interface_ref_definition: true

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."