1. Allow configuration of users/aaa through native model 2. Add port 57400 for gNMI connection 3. add ssh.Password() auth method 4. Deviation for configuring users/aaa (deviation_set_native_user)
This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.