RT-1.1: Updating check for last-notification-error-code from gnmi.get to gnmi.watch (#1854)
* RT-1.1: Updating check for last-notification-error-code from Get to Watch

After bgp session disconnect due to cease notification, last-notification-error-code
is not updated intermittently at the same time.

Upating the last-notification-error-code check from gnmi.get to gnmi.watch.await to achieve consistency.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* Update base_bgp_session_parameters_test.go

---------

Co-authored-by: Sergey Fomin <sergey.fomin@nokia.com>