DP-1.2: Adding test support for Nokia
Following changes are made in this PR -
   1. Adding testNokiaClassifierConfig and testNokiaSchedulerPoliciesConfig to cover script scenario for Nokia
   2. Adding function (SetForwardingGroupWithFabricPriority) to configure ForwardingGroup with Fabric-priority in qoscfg.go

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."