Changed testbed file for gNMI-1.14
Test uses 2 ports. This allows run this test to run without explicitly calling the `--testbed` (to point to a 2port testbed file) flag.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."