refactor: gnmi-1.12 nokia deviation for native yang instead of cli in multi origin (#2277)
* refactor: gnmi-1.12 nokia native instead of cli because its the same for srlinux

* restore metadata files

This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.

* chore: dont use deprecated GetConfig

* fix: missing pointer for generic type

---------

Co-authored-by: Sergey Fomin <sergey.fomin@nokia.com>