TE-2.2: moved static route under deviation condition (#2415)
* moved static route under deviation condition

Currently scriptapply static route override for all vendor while adding
interface+MAC NextHop instead of restrict to specific deviations.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* correcting if check

---------

Co-authored-by: Sergey Fomin <sergey.fomin@nokia.com>