TE-15.1: Adding flag to skip ipv6 compliance cases
Based on discussion in b/318469590 making following changes -
    1. Adding skip_ipv6 flag to skip recently added ipv6 tests in gribigo compliance library.
    2. Updating gribigo package version in go.mod

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."