gNOI-4.1 Adding Nokia deviation (#2462)
* gNOI-4.1 Adding Nokia deviation

Added devations to accomodate for this test passing.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* suggested fixes & merged upstream to branch

* actually adding the fixes to move ocPortName var inside if statement

---------

Co-authored-by: Sergey Fomin <sergey.fomin@nokia.com>