Enhance debug capabilities for gNOI-3.1 (#2397)
Added a comparison of the components pre/post reboot which will be
output as Logf() in the event of a failure
refactored failure message to reflect the failure more accurately
removed the uint64 casting of the time (float64) for more accurate
time checking
Reference: b/310535585

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

Co-authored-by: Sergey Fomin <sergey.fomin@nokia.com>