updated scale number and removed traffic check for failed route due to (#2640)
OTG issue #2639
Also added ARP check to make sure ARP is resolved before Traffic is
sent.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."