Update README.md
Removed ACCEPT_ROUTE for asp-policy-v4 as the expectation of the test is to validate both asp-policy-v4 and med-policy-v4 gets applied. If ACCEPT_ROUTE is configured in asp-policy-v4, route will be accepted on applying this policy and the next policy configured in export 'med-policy-v4' will not be applied. 

Verification:
Validate that the prefix ```ipv4-network-2``` i.e. ```192.168.20.0/24``` on ATE from BGP neighbor on DUT Port-1 has AS-PATH with the ASN of DUT occuring twice
    *   /network-instances/network-instance/protocols/protocol/bgp/rib/attr-sets/attr-set/as-path/as-segment/state/member
*   Validate that the prefix ```ipv4-network-2``` i.e. ```192.168.20.0/24``` on ATE from BGP neighbor on DUT Port-1 has MED set to ```1000```
    *   /network-instances/network-instance/protocols/protocol/bgp/rib/attr-sets/attr-set/state/med