RT-7.8: BGP Community Match Action Initial Commit 1. Added test script bgp_comm_match_action_test.go 2. Added new deviation bgp_actions_set_community_method_unsupported
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."