RT-2.12: Fix static to isis redistribution test
There were some of test issues:
1. The test was doing an update of the table connections policy instead of replace.
This was causing the policies to get appened to the `import-policy`, which causes
things to fail on implementations that support only one policy at a point of
attachment. Since the intent of the test isn't to validate multiple policies
at the point of redistribution, change `isisImportPolicyConfig` to use `BatchReplace`

2. The prefixes in the `prefix-set` were missing the prefix portion, so was
getting rejected by implementations that check the format of the prefix. Fix it by
adding the prefix portion to `rplPrefixMatch`

3. Fix some typos in the test descriptions

Additionally add an arista eos specific deviation `routing_policy_tag_set_embedded` since
it does not natively support tag-sets that can be referenced.