Revise use of experimental folder and update OTG references (#3030)
The experimental folder is over-used and contains many tests which in fact do pass one or more DUT's.  The previous practice encouraged use of the experimental folder by default, for example, if only a README was created, but without automation, it was unknown if the test could pass.

Instead the default should be to use the regular feature folder for new README documentation without test automation.  The experimental folder should only be used if the test automation is written and proven not to pass on any known DUT.  The intent is this is a small hint to people that the test automation is not proven to work and therefore suspect to changes.

The problem with this approach is we typically write a README before the automation is implemented.  Later when the automation is added, even if passing 1+ DUT, the folder is not  updated.