-Removing 'ExplicitPortSpeed' deviation from below IDs " RT-1.3 -BGP Route Propagation RT-1.51 -BGP multipath ECMP RT-1.52 -BGP multipath UCMP support with Link Bandwidth Community RT-1.32 -BGP policy actions - MED RT-7.8 -BGP Policy Match Standard Community and Add Community Import/Export Policy RT-7.1 -BGP default policies "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."