-Removing 'explicit_port_speed' deviation from below IDs " RT-1.33 -BGP Policy with prefix-set matching RT-1.2 -BGP Policy & Route Installation RT-1.5 -BGP Prefix Limit RT-1.27 -Static route to BGP redistribution RT-1.7 -Local BGP Test RT-1.10 -BGP Keepalive and HoldTimer Configuration Test "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."