-Removing 'explicit_port_speed' deviation from below IDs " TE-11.1 -Backup NHG: Single NH RT-1.1 -Base BGP Session Parameters TE-11.3 -Backup NHG: Actions TE-11.31 -Backup NHG: Actions with PBF TE-8.1 -DUT Daemon Failure RT-1.14 -BGP Long-Lived Graceful Restart RT-1.24 -BGP 2-Byte and 4-Byte ASN support with policy "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."