-Removing 'explicit_port_speed' deviation from below IDs " TE-9.1 -FIB FAILURE DUE TO HARDWARE RESOURCE EXHAUST TE-13.1 -gRIBI route ADD during Failover TE-13.2 -gRIBI route DELETE during Failover TE-17.1 -VRF selection policy driven TE RT-5.6 -Interface Loopback mode TE-1.2 -My Station MAC "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."