-Removing 'explicit_port_speed' deviation from below IDs " gNMI-1.11 -Telemetry: Interface Packet Counters gNMI-1.5 -Telemetry: Port Speed Test RT-2.1 -Base IS-IS Process and Adjacencies gNMI-1.10 -Telemetry: Basic Check "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."