-Removing 'explicit_port_speed' deviation from below IDs " TE-3.6 -ACK in the Presence of Other Routes TE-11.21 -Backup NHG: Multiple NH with PBF TE-11.2 -Backup NHG: Multiple NH TE-4.1 -Base Leader Election TE-3.7 -Base Hierarchical NHG Update TE-3.1 -Base Hierarchical Route Installation "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."