-Removing 'explicit_port_speed' deviation from below IDs " TE-16.2 -encapsulation FRR scenarios TE-5.1 -gRIBI Get RPC TE-15.1 -gRIBI Compliance TE-14.1 -gRIBI Scaling TE-14.2 -encap and decap scale "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."