-Removing 'explicit_port_speed' deviation from below IDs " TE-2.1 -gRIBI IPv4 Entry TE-2.2 -gRIBI IPv4 Entry With Aggregate Ports TE-3.5 -Ordering: ACK Received TE-4.2 -Persistence Mode TE-6.2 -Route Removal In Non Default VRF "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."