-Removing 'explicit_port_speed' deviation from below IDs " TE-6.1 -Route Removal via Flush TE-8.2 -Supervisor Failure TE-3.2 -Traffic Balancing According to Weights RT-5.4 -Aggregate Forwarding Viable RT-5.2 -Aggregate Interfaces "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."