-Removing 'explicit_port_speed' deviation from below IDs " RT-5.3 -Aggregate Balancing RT-5.1 -Singleton Interface TE-1.1 -Static ARP MTU-1.3 -Large IP Packet Transmission OC-1.2 -Default Address Families P4RT-1.2 -P4RT Daemon Failure gNMI-1.18 -gNMI subscribe with sample mode for backplane capacity counters gNMI-1.13 -Optics Telemetry "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."