-Removing 'explicit_port_speed' deviation from below IDs " DP-1.11 -Bursty traffic test DP-1.10 -Mixed strict priority and WRR traffic test DP-1.7 -One strict priority queue traffic test DP-1.14 -DSCP transperency with ECN DP-1.4 -QoS Interface Output Queue Counters DP-1.8 -Two strict priority queue traffic test DP-1.9 -WRR traffic test SFLOW-1 -sflow and sFlow sample data "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."