-Removing 'ExplicitPortSpeed' deviation from below IDs " RT-1.12 - BGP always compare MED RT-7.5 - BGP Policy - Match and Set Link Bandwidth Community RT-1.26 - Basic static route support RT-2.13 - Weighted-ECMP for IS-IS gNMI-1.21 - Integrated Circuit Hardware Resource Utilization Test TRANSCEIVER-4 - Telemetry: 400ZR RX input and TX output power telemetry values streaming RT-1.34 - BGP route-distance configuration RT-1.30 - BGP nested import/export policy attachment RT-1.29 - BGP chained import/export policy attachment "
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."