-RT-1.23 following changes incorporated.
-Changes in metadata.textproto

  -explicit_interface_in_default_vrf
  -bgp_global_extended_next_hop_encoding_unsupported
  -skip_bgp_session_check_without_afisafi

-Changes in bgp_afi_safi_defaults_test.go

  -Removed unused ports from verifyPortsUp func.
  -Enabled global AFI-SAFI configuration.
  -Unsetting IPV4 AFI-SAFI flag under nieghbour level
  -Added deviation ExplicitInterfaceInDefaultVRF, BGPGlobalExtendedNextHopEncodingUnsupported

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."