RT-1.34 admin_distance_test changes incorporated
-Changes in metadata.textproto

-vender check added for NOKIA
-Added deviations explicit_interface_in_default_vrf & interface_enabled

-Changes in admin_distance_test.go

-Enabled "configureRoutePolicy" configuration for ebgp.
-Added interface configuration for port3

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."