fixed the failure that was introduced by PR-3376
PR- 3376 changes changed the signal type to KILL instead of TERM  that
resulted to non graceful application restart , fixed by adding that as an parameter to the  KillProces() function.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."