Dropoctets validation to be inline with droppkts
1. Modified the Drop octets validation to be inline with dropped pkts.
   It was validating based on OTG packets diff and packet size , instead
changed to check for 0 as drops are not expected to be seen.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia’s intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."