- RT-5.9 disable_ipv6_nd_ra_test changes incorporated
- Updated metadata.textproto
  -vender check added for NOKIA
  -Added deviations explicit_interface_in_default_vrf & interface_enabled

- Updated disable_ipv6_nd_ra_test.go
  - Enabled ExplicitInterfaceInDefaultVRF check
  - Removed set mode RouterAdvertisement_Mode_ALL.

“Removed the ‘set mode All’ configuration for IPv6 Router Advertisements due to issue #2981.

Refer issue - https://github.com/openconfig/featureprofiles/issues/2981

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."