RT-1.27 : Fixing script issues seen after recent changes
Following changes are made -
  1. Fixing the usage of 'deviations.SetMetricAsPreference', i.e. when this deviation is set true, script should configure the Preference with Metric value.
  2. Restoring the original static route as per steps in readme, which make traffic validation possible.
  3. Fixing 'validateLearnedIPv4Prefix' and 'validateLearnedIPv6Prefix' to validate rejected-route when 'shouldBePresent' is set to false.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."