RT-1.32: BGP policy actions - MED LocPref incorporated
Updated metadata.textproto
- ExplicitPortSpeed removed.

Updated actions_MED_LocPref_prepend_flow_control_test.go
- The check for `DefaultRoutePolicyUnsupported` in the Go code has been modified, leading to an unintended deviation that affects feature supported vendor.
- The log messages have been modified to print proper value.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."