RT-1.29: BGP chained import/export policy attachment changes incorporated
Updated metadata.textproto

 vender check included for NOKIA
 Added deviations.

Updated chained_policies_test.go

 Removed accept route for statement match-policy-v4 & match-statement-v6 as the next statement prefix-set-v4 & prefix-set-v6 will not be applied if accept route is configured.
 Refer PR- #2760

 Added deviations for prefixset_Mode.

 Updated to use batch replace for delete routing-policies.

Updated README.md

 policy-result as `ACCEPT_ROUTE` not required for match-policy-v4 & match-statement-v6 policy.
 Refer PR- #2760

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."