RT-2.12 Fixed with deviations & minor changes - Edited dp2v6Route such that it is advertised as ISIS route. - Set separate tag value for IPv4 & IPv6 as per README. - Added missing metricPropogation parameter.
PS: Script needs alteration (apart from this PR) such that it adheres to README.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."