adding waitForARP to fix flakiness
Traffic is sent without waiting for ARP and that is causing for flakiness
in result , this change will make sure ARP is resolved

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."