RT-1.30: BGP nested import/export policy attachment changes incorporated
Updated metadata.textproto

  vender check included for NOKIA
  Added deviations.

Updated nested_policies_test.go

  Updated the code as per the instructions in README.md, replacing ACCEPT_ROUTE with NEXT_STATEMENT.
  Added the definitive statement for NEXT_STATEMENT.
  Added deviation for prefixset_Mode.
  Added deviation to skip rp match set options.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."