Added defer function to cleanup fib table (#3561)
currently after test is complete OTG retains its config and due to which
following tests are failing as fib is full and not able to install any
more routes.  added defer function to stop BGP protocol that will clear
all routes injecected .

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."