RT-1.51: BGP multipath ECMP changes incorporated
Updated metadata.textproto

  Added deviation enable_multipath_under_afi_safi
  Removed deviation explicit_port_speed

Updated bgp_multipath_ecmp_test.go

  Added the vendor switch for Nokia, to skipped multipath configuration under peer-group.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."