RT-7.11: BGP Policy - Import/Export Policy Action Using Multiple Criteria
Updated metadata.textproto
 -vender check included for NOKIA
 -Added deviations.

Updated import_export_multi_test.go
 - Added community-set-refs configuration for add_comm 60/70.
 - Delete routePolicy policy applied to neighbour before configurating multipolicy.
 - Added deviation for prefixset_Mode.
 - Policy result changed to ACCEPT from NEXT-STATEMENT for policy match_community_regex statements.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."