README.md updated
The policy "match_community_regex" contains a single, non-nested statement with a policy result "NEXT-STATEMENT," which is non-terminating. Therefore, the policy result has been updated to "ACCEPT_ROUTE," a definitive action.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."