RT-7.2: BGP Policy Community Set changes incorporated
Updated metadata.textproto
 - Removed "bgp_conditions_match_community_set_unsupported" deviation.

Updated community_test.go

 - "match-set-options" configuration under the "defined-sets" path has been removed.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."