Updated actions_MED_LocPref_prepend_flow_control_test.go - Restoring unrequired changes - The "deleteNbr" detail has been updated for RT-1.32.5 to address an issue where an unintended neighbor was being used for deleting the import and export policy.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."