RT-7.8: BGP Policy Match Standard Community and Add Community Import/Export Policy changes incorporated (#3593)
* RT-7.8: BGP Policy Match Standard Community and Add Community Import/Export Policy changes incorporated

Updated metadata.textproto
 - Removed "bgp_conditions_match_community_set_unsupported" deviation.
 - Removed "bgp_community_set_refs_unsupported" deviation.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* - Removed "explicit_port_speed" deviation.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

---------

Co-authored-by: Tushar Rathod <111466464+trathod1@users.noreply.github.com>