SFLOW-1: Configuring sflow collector respective to IPv4 and IPv6 (#3471)
* SFLOW-1: Configuring sflow collector respective to IPv4 and IPv6

Following changes are made -
   1. Removing explicit_interface_in_default_vrf, considering test is using non-default VRF mvrf1 for the interface.
   2. Based on discussion b/290965827, refactoring the sflow collector configurations to be part of respective V4 and V6 test.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* Updating script to send 10M+ packets to have consistency with the sample
generation.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* Incorporating changes based on reveiw comment

* Updating sflow packet count with tolerance check to 8 from 9 for consistency

* Incorporating review comment

* Optimizing to avoid duplicate variables