TE-11.21:  Backup NHG: Multiple NH with PBF changes incorporated
Updated metadata.textproto
- Removed port speed devation
Updated backup_nhg_multiple_nh_pbf_test.go
- Added port-1 to default network instances

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."