RT-14.2: gRIBI route test changes incorporated.
- Updated metadata.textproto
  - Removed explicit_port_speed
  - Added no_fallback_route_in_transit_vrf

- Updated gribi_route_test.go
  - A deviation has been included to skip TC-14.2.4, as fallback route support is not available in Transit network instances.
  - Removed explicit port speed check.

- Updated Deviation.go,deviations.go & metadata.pb.go
  - A new deviation has been included due to the unavailability of fallback route support in Transit network instances.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."