RT:2.1: Enabling deviation isis_explicit_level_authentication_config
  - Enabling deviation isis_explicit_level_authentication_config
  - Updating deadline/timeout values during checks for consistent results

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

Co-authored-by: Marc Charlebois <marc.charlebois@nokia.com>