MTU-1.3: Removed deviation in large IP packet transmission.
Updated metadata.textproto
- Removed explicit_port_speed
- Removed omit_l2_mtu

Updated large_ip_packet_transmission_test.go
- Changes ensure that the omit_l2_mtu deviation applies only to L2 MTU and does not affect IP MTU configuration under the subinterface.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."