TRANSCEIVER-12 Fixed with adding required deviations (#3638)
* TRANSCEIVER-12 Fixed with adding required deviations
  -deviation `require_zr_oper_mode` --> https://partnerissuetracker.corp.google.com/issues/383369830
  -deviation `explicit_dco_config`  --> https://partnerissuetracker.corp.google.com/issues/383075189

* Eric comments incorporated:
1. operationalMode flag added.
2. InterfaceConfig function updated to use ConfigOpticalChannel.
3. ConfigOpticalChannel function updated to configure in single operation.
4. Removed RequireZrOperMode deviation.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

---------

Co-authored-by: MarcCharlebois <marc.charlebois@nokia.com>