TRANSCEIVER-4: Added InterfaceConfig & increased sleep time - Initialized OperationalModeFlag. - Added InterfaceConfig. - Increased sleep interval to 80Sec as per buganizer:384969268 https://partnerissuetracker.corp.google.com/issues/384969268
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."