TRANSCEIVER-5 Fixed with multiple changes - Added operationalMode flag & initialize for interface.go - Integrated streamOC1 & streamOC2 such that can be validated in single validation. - gnmi.Watch to take effect optical-channel parameter setting & added delay of 80Sec as per b/384969268 - Minor change in error logs.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."