TRANSCEIVER-6 Fixed with initializing operationalMode & increasing delay - Initialized operationalMode - Increased delay of 80Sec to stabilize telemetry as per b/384969268.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."