TRANSCEIVER-10 Fixed with operationalMode initialization - Initialized operationalMode with flag. - Used deviation eth_channel_ingress_parameters_unsupported in interface.go for unsupported logical-channel/ingress/transceiver leaf as per https://partnerissuetracker.corp.google.com/issues/297268414#comment26.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."