Fix P4RT-2.2 and TE-2.2 intermittent test failures for EOS (#4008)
* revert the previous change

* Fix P4RT-2.2 and TE-2.2 intermittent test failures for EOS
P4RT-2.2 was failing because the testbedType is wrong in
metadata.textproto. This is used to construct the right binding file to
use for the test
TE-2.2 needs a interlock between shut/unshut of the link and when the
traffic is started to validate the traffic.

* Revert "Fix P4RT-2.2 and TE-2.2 intermittent test failures for EOS"

This reverts commit a20cbf4cbf04c704b2b653e79750cd4b171d8f3e.

removing this commit from the main branch#

* Fix P4RT-2.2 and TE-2.2 intermittent test failures for EOS
P4RT-2.2 was failing because the testbedType is wrong in
metadata.textproto. This is used to construct the right binding file to
use for the test
TE-2.2 needs a interlock between shut/unshut of the link and when the
traffic is started to validate the traffic.

* remove inadvertent change