- In func validateImportPolicyDut added validation of 3 prefixes on OTG. - In func validateExportPolicyDut added validation of prefixes on OTG. - In validateImportRoutingPolicyAllowAll added validation of v4 & v6 prefixes on OTG.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."