TRANSCEIVER-13: Fixed by adding deviations & initializing operationalMode. - Description leaf deviation added as per https://partnerissuetracker.corp.google.com/issues/394622453 - Output power interval leaf deviation added as per https://partnerissuetracker.corp.google.com/issues/394622454 - OperationalMode initialized & metadata updated.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."