RT-7.3: Fixing IPv6 route generation, DefaultImportPolicy and regexp pattern (#3632)
* RT-7.3: Fixing IPv6 route generation, DefaultImportPolicy and regexp pattern

Following changes are made here -
  1. Fixing IPv6 route generation to have AsPath set as expected by Readme test and as it was done for IPv4.
  2. Fixing DefaultImportPolicy for IPv4 to "Reject" as required by eBGP, and as it was before PR-3025
  3. Removing ^ from the regexp aspathMatch from the beginning, considering these eBGP routes will always AsPath with neighbor AS "65511" at the start, and will never match "^100"

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* Fixing failure seen with vendor having CommunityMemberRegexUnsupported