RT-5.5 Fix to ensure lag interface operational state & infra changes. (#4088)
* RT-5.5 Fix to ensure lag interface operational state & infra changes.
- Ensure lag interface operational state UP in both OTG & DUT before proceeding testcases.
- Ensure lag interface operational state Down before proceed TC3 verification.
- Fix infra changes.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* -Added 4Sec delay as per README in TC3ShortUP test.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."