TRANSCEIVER-11: Initialized operationalMode & OTN to ETH channel config added (#4007)
* TRANSCEIVER-11: Initialized operationalMode
- Initialized operationalMode with flag.
- Used deviation `eth_channel_ingress_parameters_unsupported` in interface.go for unsupported logical-channel/ingress/transceiver leaf as per https://partnerissuetracker.corp.google.com/issues/297268414#comment26.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* -Removed unused variable & packages.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* OTN to Eth logical channel configuration in ConfigOTNChannel func as per https://partnerissuetracker.corp.google.com/issues/384673249

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."