RT-2.12: Replaced GetAll with WatchAll & fixed watchAll conditions (#4180)
* rt-2.12: Replaced GetAll with WatchAll & ensured ISIS adjacency.
  - Replaced GetAll with WatchAll for Metric propagation sub-cases in both v4 & v6.
  - Ensuring ISIS adjacency during every sub-case execution.
  - Minor Log message fixes.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* - Restored ISIS back to original & ensured ISIS adjacency before testing sub-cases.
- Fixed verifyPrefix function for 2.12.5 & 2.12.7 sub-cases.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."