TE-17.1: Fixed ttl,vrf_c,gribi programming - Fixed incorrect ttl expectation post decap. - Fixed vrf_selection_policy_c rule15 & 16. - Set isis global afi-safi for ipv6. - Appended subinterface setting .0. - Fixed gribi NHG & ipv4 entry programming.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."