RT-2.13 Adding MaxECMP Paths and Deviations (#4124)
* * (M) feature/isis/otg_tests/weighted_ecmp_test/metadata.textproto
 - Added required Deviations
* (M) feature/isis/otg_tests/weighted_ecmp_test/weighted_ecmp_test.go
 - Removed unrequired ExplicitInterfaceInDefaultVRF in ISIS code block and added that deviation to handle LAG/Interface
 - Added MaxECMP paths to be 3 (as many as the number of Lags)
* (M) feature/isis/otg_tests/weighted_ecmp_test/README.md
 - Added MaxECMP in the ConfigPaths

---

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* ran gofmt

---

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."