RT-7.8: comm_match_action_test changes incorporated.
Updated bgp_comm_match_action_test.go

- Removed SendCommunityType from the peer-group configuration and added it under AFI-SAFI configuration.

Updated metadata.textproto

- Removed "skip_bgp_send_community_type" deviation.

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."