ExRR README.md (#3474)
* Create README.md

* ERR README.md

This is to define a new Test for ERR (Extended Route retention). This test demonstrates a new behavior for Graceful restart which is an extension to the existing behavior explained in RFC4724 and RFC8538.

This test also relies on,

- Changes proposed to the gNOI.bgp proto in https://github.com/openconfig/gnoi/pull/214
- Also relies on some OC paths for ERR which arent available yet.

* Update README.md

Formatting changes

* Update README.md

Updated the Yaml data

* Update README.md

Updated to 80 Characters

* Update README.md

* Update README.md

Added ERR policy details under summary

* Update README.md

Updated the Yaml definition to include OC paths for config and STATE

* Update README.md

* Update README.md

Adding ":" to paths defined under Yaml

* Update README.md

Update Yaml file to run the checks successfully. Path for "config/reject-route" had duplicates.

* Update README.md

* Update README.md

Some more corrections to Yaml paths and rpc

* Update README.md

More corrections to the YAML section

* Update README.md

* Update README.md

Formatting

* Update README.md

Significant Modifications to test. Added several new sub-tests

* Update README.md

Indentation corrections

* Update README.md

Fixing tables

* Update README.md

Correction to the Markdown tables

* Update README.md

Correction to the Markdown tables

* Update README.md

Updates to the test to make RT-1.35.2 a test for DUT as helper

* Update README.md

ERR --> ExRR

* Update README.md

Commenting the OC-paths for ExRR.

* Update README.md

removed the currently unsupported OC paths for Exrr

* Update README.md

Updates to RT-1.35.6 clarifying that the intent of the test is to ensure the DUT
1. Generates the correct notification message as per https://github.com/openconfig/gnoi/pull/214
2. The DUT flushes all the routes on receipt of the gNOI.BGP HARD_RESET command

* Rename README.md to README.md

Changed file name

* Rename README.md to README.md

reverting back the file name

* Update README.md

changes to the highest value supported

* Update README.md

* Rename README.md to README.md

Change folder name from ERR to EXRR

* Rename README.md to README.md

Changing the file path to comply with the checks

---------

Co-authored-by: rszarecki <46606165+rszarecki@users.noreply.github.com>
Co-authored-by: Ram <rmachat@google.com>