RT-1.27: Enabling default_bgp_instance_name and fixing recent changes (#4342)
* RT-1.27: Enabling deviation default_bgp_instance_name

Updated metadadta.textproto with default_bgp_instance_name: "BGP"

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* Fixing failures introduced after recent script changes