RT-2.8 Added required deviations & delay to fix Added isis_multi_topology_unsupported as multi-topology is not supported. Added isis_dis_sysid_unsupported DisSystemId not supported. Added isis_global_authentication_not_required as not required. Added isis_explicit_level_authentication_config for level authentication. Updated missing_value_for_defaults config in .go for system level counters.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."