TE-14.2 fix logics & added repaired_vrf - Added vrf_repaired as vrf_r config without vrf_rd is not logical. - Removed ! with deviations.DecapNHWithNextHopNIUnsupported. - Under deviations.ExplicitInterfaceInDefaultVRF(dut) included DefaultVRFIPv4NHCount while configuring NHs. - Under BuildVRFCfonfig function added missing code which was included in TE-14.1. - For Nokia vendor added decapIPv4ScaleCount=128. - Added `override_default_nh_scale` deviation. - Existing scale.go configures 29254 NHs where as many NHs are unused so changed logic to configure NHs which are needed(9511 NHs).
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."