CNTR-2 Fixes
(m) cntr_test.go
- add function to wait for containerReadiness to avoid repeated code
- add deviation to push interfaces to default vrf
- add wait for containerReadiness for final subtest

(m) cntrsrv.go
- let client gracefully handle EOF for empty getResponse

---
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."