RT-5.7: Aggregate Not Viable All changes incorporated.
- Introduced a deviation so the script uses interface names and MAC next-hops at the first-level next-hops.
- Removed vendor specific load-balancing checks. Vendors are expected to manage unsupported values using deviations. Refer the PR comment for details.
  https://github.com/openconfig/featureprofiles/pull/3550/files/fdab107be3d5b0c617e386d513d33dfb85a5f73a

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."