Register FNT for Auto Link-Bandwidth Community (#4680)
* Update testregistry.textproto

Register FNT for Auto Link-Bandwidth Community

This commit pre-allocates a new test ID and path in the test 
for the forthcoming Functional Network Test (FNT) covering the
BGP auto-generated link-bandwidth feature.

This is a preparatory change. The full test specification (README.md)
will be added in a subsequent commit.

* Update testregistry.textproto

add exec