TRANSCEIVER-101,102 fix - Added 800zr testbed in runtests.go - Removed unwanted Nokia deviation `eth_channel_ingress_parameters_unsupported`. - As per README changed functions under zr_validation.go   i. Configured oper-mode, frequency & targetOpticalPower with admin-state Up.  ii. Removed redundant interface admin-state bouncing.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."