RT1dot12: updated the script with gNMI watch to verify received MED (#3866)
* update the scriptwith gNMI watch to verify recieved MED

1. added watch to verify MED value for recieved prefixes
2. restricted the port status check to the ports that has OTG config
3. added deviation for bgp instance name
4. Removed port speed deviation

"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."

* updated test as per comments