TRANSCEIVER-5.2: Restructured using itemized functions from internal library -Defined `TunableParamsTest` which validates optical-channel telemetry. -Added 3 more conditions in `isOpticalChannelStreamReady` func. -Added required deviations in metadata.textproto. -Used `tunableParamsTest` func to restructure zrp_tunable_parameters_test.
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."